Introduce ProcessedDatasetImporterService for clarity
